AI structure change, AI movement fixed, AI prefab, Spawner Prefab
I replaced the ai structure, the decision tree wasn't really necessary and too complex for what we needed. Changed it to be a self-contained file.   Ai will slow down to shoot and attack the target, changing course if the player is within a certain distance(1500 atm).  Spawners will spawn a modifiable number of units within the boxcollider component they possess.  Enemy unit has  a prefab with an ai script, the correct orientation, and movement.  The spawner also has a prefab.